requirements.txt
# ── Email Protocols ────────────────────────────────────────
# imaplib, poplib, smtplib → built-in Python stdlib (no install needed)

# ── Gmail API ─────────────────────────────────────────────
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# ── Microsoft Graph / MSAL ────────────────────────────────
msal>=1.24.0

# ── HTTP & API ────────────────────────────────────────────
requests>=2.31.0
aiohttp>=3.9.0

# ── URL Analysis ──────────────────────────────────────────
tldextract>=5.0.0
validators>=0.22.0
dnspython>=2.4.0
python-whois>=0.8.0

# ── NLP / ML ──────────────────────────────────────────────
scikit-learn>=1.3.0
transformers>=4.35.0
torch>=2.1.0                  # OR: tensorflow>=2.14.0
spacy>=3.7.0
nltk>=3.8.0
# After install, run:
#   python -m spacy download en_core_web_sm
#   python -c "import nltk; nltk.download('stopwords'); nltk.download('punkt')"

# ── File Type Detection ───────────────────────────────────
python-magic>=0.4.27
# On Ubuntu/Debian: sudo apt install libmagic1
# On macOS:         brew install libmagic

# ── Malware Scanning (ClamAV) ─────────────────────────────
pyclamd>=0.4.1
# Install ClamAV: sudo apt install clamav clamav-daemon && sudo freshclam

# ── Utilities ─────────────────────────────────────────────
pyyaml>=6.0.1
tqdm>=4.66.0
colorama>=0.4.6

# ── (Optional) PDF / XLSX processing ─────────────────────
# pdfplumber>=0.10.0
# openpyxl>=3.1.0
